+++ docs/Changelog.md
@@ -5,6 +5,7 @@
 - Android: scoped storage example [#785](https://github.com/mapsforge/vtm/pull/785)
 - Mapsforge: map stream support [#784](https://github.com/mapsforge/vtm/pull/784)
 - Render theme from Android content providers [#783](https://github.com/mapsforge/vtm/pull/783)
+- Render theme xml pull parser [#786](https://github.com/mapsforge/vtm/pull/786)
 - Many other minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.15.0)